Move swap integration test
